Struct polyplets::SecurityTetraplet

pub struct SecurityTetraplet {
    pub peer_pk: String,
    pub service_id: String,
    pub function_name: String,
    pub lens: String,
}

Describes an origin that set corresponding value.

Fields
peer_pk: String

Id of a peer where corresponding value was set.

service_id: String

Id of a service that set corresponding value.

function_name: String

Name of a function that returned corresponding value.

lens: String

Value was produced by applying this lens to the output from call_service.

Implementations

impl SecurityTetraplet

pub fn new( peer_pk: impl Into<String>, service_id: impl Into<String>, function_name: impl Into<String>, lens: impl Into<String>, ) -> SecurityTetraplet

pub fn literal_tetraplet(init_peer_id: impl Into<String>) -> SecurityTetraplet

Create a tetraplet for string literals defined in the script such as variable here (call ("" "") "" ["variable_1"]).

pub fn add_lens(&mut self, lens: &str)